requirements.txt
streamlit
langchain
langchain-groq
langchain-community
python-docx
python-dotenv
pdfplumber